requirements.txt
numpy
pyzmq
PyTurboJPEG
Pillow